Fix weapon system: change magSize to int, update ammo UI
- Changed magSize from float to int in WeaponSO (magazine sizes should be whole numbers)
- Updated scene UI references: renamed ammo display GameObjects and added totalAmmo reference
- Updated weapon SO references to use new weapon configuration
- Fixed ToString() call to work correctly with int magSize type

+++ Assets/ScriptableObjcts/WeaponSO.cs
@@ -5,7 +5,7 @@ public class WeaponSO : ScriptableObject
 {
     public int damage;
     public float firerate;
-    public float magSize;
+    public int magSize;
     public GameObject weaponPrefab;
     public bool canZoom;
     public float zoomFov;
